SSL.File: Move read_callback setting to read().
The stream read_callback is now set (to ssl_read_callback) whenever
the read buffer isn't full, and cleared when it is.

This reduces the complexity and cleans up the semantics of where
and when the stream read_callback is changed.

Also makes similar changes to the handling of the stream close_callback.

This means that the function update_internal_state() is now a noop.