SSL.File: Changed backend handling.
The backend to use (real/local) now strictly depends on blocking/
nonblocking mode. This means that creation of the local backend
can be (and has been) moved to set_blocking{,_keep_callbacks}().

This reduces the number of short-lived backends used by SSL in a
typical nonblocking backend-driven application down to ~ zero.