Override write and werror in hilfe
The overriden version uses readline->write to actually print the data.

This is rather convenient when writing asyncronous code in hilfe,
newlines are handled correctly and it interacts fairly well with the
input system.

The overriding is intentionally using add_constant to also override
debug output in modules that are loaded after hilfe is started.