Debug: Changed trace flag for check_call_svalue().
This moves tracing of check_call_svalue() from using l_flag (aka
optimizer_debug()) to using trace_level (aka predef::trace()).

The rationale for this is that check_call_svalue() is used by the
runtime (via the variant dispatcher), and not by the compiler.

Tracing of check_call_svalue() is (as before) enabled at level 3.