[precompiler] Use the correct types for classes and objects
This includes programs in the cmod file and external programs.

This makes the typechecker less odd, as an example the IOBuffer class
functions that return an IOBuffer will not give warnings along the lines
of "object of type IOBuffer can not be returned from a function
returning object(IOBuffer=__Stdio()->IOBuffer)".

You can also use any object type as arguments and return types, although
you still have to use object(X.Y) in most places due to the parser.