Nettle: Reference the programs being inherited.
Make sure that there's a local symbol in the parent for the
programs that we inherit. Otherwise the parent_identifier in
the generated inherit may point back to some later override.

This happens frequently when inheriting the State class.

This fixes infinite recursion in eg Program.all_inherits()
when called with several of the Crypto and Nettle classes.