set_priority(): Fix going from realtime to normal priorities.
The following code sequence (and many others) used to fail on systems
that have both the sched_setscheduler() and setpriority() syscalls:

    set_priority("realtime");
    set_priority("lower");

The first call would set the policy to SCHED_FIFO.  The second call
would not touch the policy, but just use setpriority().  The result is
that the process continues to run with realtime priority.

Fix by ensuring that the code calls both sched_setscheduler() and
setpriority() when SCHED_OTHER is in effect.

At the same time, reorganize the code somewhat to be more readable.

(cherry picked from commit 48aa6ad98e390fb68e2055fa67f951e5fb32159f)