Two somewhat related fixes to the PNG decoder
1: Handle any number of PNG chunks (up to the available memory)

2: Stop decoding once the IEND chunk has been processed.

This fixes opera bug TURBO2-767:
   Crash when decoding http://extrafahaz.hu/fotok/image052e.png

The image in the bug has a lot of data appended after the IEND chunk,
and in total after decoding there was 339017 chunks.

Since the decoder used the stack to build the array this meant that it
had a tendency ran out of stack space, and caused fairly random
crashes.

Fixed by simply avoiding the stack, instead an array is built
directly.

Also, the PNG format says that the file ends after the IEND chunk, the
decoder now enforces that, which makes the file decode significantly
faster (now ~60 chunks instead of 339k).