Build: Improved support for --disable-binary.
Don't attempt to make depend in post_modules when cross-compiling
or --disable-binary, since a fresh pike is needed to generate the
sources there.

Fixes issues with source dists generated with --disable-binary
and an old system pike, where broken files would show up in
the resulting archive and not be regenerated on actual build.