Testsuite: Test cipher-suite driven downgrade.
The full suite is now run when the minimum protocol version
is SSL 3.0 for both client and server, and the maximum
protocol version is the same for client and server.

This allows testing of downgrading to older protocol
versions due to the only common cipher suite being
removed in the common protocol version.

Also shortens "expected" to "exp" in lots of places to get past the
m4 4096 byte limit.