Compiler: Added #pragma {,no}disassemble.
When debugging the compiler it is is often useful to get
disassembly output for just a few functions in a specific
file. This pragma causes the compiler backend to dump the
result from the peep-hole optimizer to stderr for functions
defined in sections where #pragma disassemble is active.

Note that the pragma currently only has effect if pike
has been compiled --with-rtldebug (aka PIKE_DEBUG).