SSL.ClientConnection: Support TLS 1.3 ClientKeyShare.
client_hello() now sends a pure TLS 1.3 hello if Context.min_version
is TLS 1.3 or later, a compat TLS 1.3 hello if Context.max_version is
TLS 1.3 or later, and a legacy TLS 1.2 or earlier handshake otherwise.

Note that SSL.ServerConnection does not yet suport pure TLS 1.3 hellos,
and that the SSL.ClientConnection won't be happy with the result from a
server that does support TLS 1.3.