SSL.ClientConnection: Improve interoperability with stupid servers.
It seems there are many SSL/TLS servers out there that refuse packets
that have a higher version than the TLS version that they support.

There are also TLS servers that refuse packets that have version SSL 3.0.

Note that all of the above servers are in violation of RFC 5246 E.1:

  Thus, TLS servers compliant with this specification MUST accept any
  value {03,XX} as the record layer version number for ClientHello.

All ClientHello packets are now sent with packet version TLS 1.0 except
ClientHello SSL 3.0 which is sent with packet version SSL 3.0. All other
packets are sent with the negotiated version of SSL/TLS.

This should improve interoperability with the affected servers somewhat.