Testsuite: added testcase for undefined variables in sscanf
The value of variables declared in sscanf calls is undefined if the
sscanf call does not actually happen. It is impossible to detect this at
compile time, so generating an error in cases where they are incorrectly
used is not an option. The behavior proposed by this testcase is to
always initialize them to UNDEFINED.

Thanks to Chris Angelico for bringing this issue up and providing a
testcase.