Filesystem.Tar: Survive existing paths with restricted permissions.
extract() now attempts to temporarily alter permissions for
existing directories and files if they are likely to cause
the extraction to fail.

The typical case is a write-protected directory or file in a tar
file which is extracted on top of an earlier extraction. This
earlier caused extract() to fail with "Failed to create \"xx\":
Permission denied." or similar.

Fixes [bug 7350].