SSL.Context: Reduce aggressivity of purge_session() for <= TLS 1.2.
In TLS 1.2 and earlier it is possible to have multiple concurrent
connections using the same session. In particular there may be a
concurrent connection performing session resumption handshaking
at the same time as the session is being purged.

Fixes "Internal server error: Bad argument 1 to sizeof()." in
__builtin.Nettle.Hash() called via Connection.hash_messages(),
which was often triggered by Google Chrome.