SSL: Improved robustness of the async_tls_close test.
The test is run four times in succession, and attempted to use
the same (fixed) port number every time. This would cause the
test to fail if something else already had claimed the port
(unlikely), or if the OS had a too long grace period (common).

The test now lets the OS allocate a port.

Also adds a timeout if the TLS connection doesn't wake up when
the port is closed.

This should fix intermittent testsuite failures on several platforms,
where the testsuite watchdog killed the SSL testsuite.