Fixed circular reference in Protocols.HTTP.Server.Request
The OutputBuffer class has references to the parent object, i.e., the
request object. This creates a circular dependency that is not cleared
without a gc().

This meant, especially using the old Stdio.Buffer code that never
actually shrunk the size of the buffer, that a webserver using this
code would apear to leak a lot of data, especially when sending large
replies as strings, buffers for all the data sent would stay around
until the next gc(), which can be a rather long time.