Experiment with replacing the shared string hashtable code
The original reason was to make the string struct smaller´

This uses Robin Hood style hash (see also, google).

It seems to be noticeably slower, but saves memory since it does not
need the next pointer. Also removed the hval value in the strings
(this was perhaps a bad idea, the point was to speed up lookups with
linearly stored hashvalue comparisons before looking in the string
structure, but it slows down removal of strings, and actually wastes a
small amount of memory since the hashtable size is bigger than the
string table size)

In theory it should actually be just as fast due to a rather linear
access pattern, but this is sort of broken in pike since to compare
the values in the keys we need to dig around in the fairly randomly
located pike_string structures.

The solution to that might be to have pike_string structures in the
array instead of pike_string *, copying the checked fields from the
string to the hashtable struct (aka: hval, size_shift, len, str).

But the additional memory usage will probably slow things down
instead.