Initial substring support.
This adds support for substrings, and conversion of said strings to
static strings.

Left to do: gc conversion of substrings to malloced/short strings.

This requires two new stringtypes, at least.

The reason is that the substring string type is bigger than the normal
strings. I will most likely do it by adding a 2 or 1 to the 'parent'
pointer.