Build: Improved forward compatibility with Pike 8.0 precompiler.
The precompiler in Pike 8.0 and later defaults to declaring PIKECLASS
programs static. This conflicts with Pike 7.8's header files which
declares several of the programs extern.

Also adds a work-around for some versions of the Pike 8.0 precompiler
where the guard for using set_program_id_to_id() had false positives.