Filesystem.Monitor: Survive an empty monitor_queue in check().
Preparation for only having polled monitors in the monitor_queue.

Currently monitors and monitor_queue always have the same size,
but with filesystem event systems there's no need to poll the
event monitored files by hand, so leave the monitor queue for
the paths that the event system doesn't support.