Filesystem.Monitor [Inotify]: Use the watchdescriptor to look up the monitor.
The path in an inotify event is usually just a single path segment,
and can typically not be used to look up the corresponding monitor.

This reduces the number of cases where a full monitor scan is performed
to negligible.

Also adds some handling of the IN_IGNORED event, which indicates that
a watchdescriptor has become invalid.