On this branch portable bytecode generation is not normally done
This actually noticeably speeds up the compiler and also saves about
15% of RAM when starting hilfe.

If the decode_value code is modified to free the portable bytecode
_TWICE_.

This will not be merged to 8.1 until I figure out _why_ that is. It
should only need to be freed once, but I guess it's in the string
table twice?