Odbc: Detect width of characters dynamically.
It seems unixODBC 2.3.x is confused about the width of wide characters.
In some circumstances (eg with freetds 0.91), wide fields are returned
in UCS-4, but queries and field labels in UTF-16. The header files
switch between UTF-16 and UCS-4 for SQLWCHAR with SQL_WCHART_CONVERT,
but forget to remap the W-functions accordingly.