Filesystem.Monitor [Inotify]: Reduce the number of complaints.
The inotify kernel subsystem has a maximum number of concurrent
watchpoints (controlled via /proc/sys/fs/inotify/max_user_watches).
With the default at just 8192 it is not hard to run out of watchpoints,
at which point the subsystem will start to complain.

With this change only every ~100th failure of this type will be logged
to stderr.