Protocols.HTTP.Query: correctly handle response length
According to the RFC, the content-length header field ought to be
used only if the status code is not 1xx, 204 or 304; the response
is not to a HEAD request and the transfer encoding is identity.
This is a first attempt to improve the handling of some cases, in
particular it lets Query correctly handle the special status codes.