Runtime [get_memory_object_memory]: Always set the shift on success.
The shift value needs to be set when not returning MEMOBJ_NONE.

Previously it was left unmodified in in the narrow
MEMOBJ_STRING_BUFFER case.

Fixes SIGSEGVs in the MsgPack testsuite on at least Solaris 8/sparcv9,
and it has been pure luck that has seemed to work on other platforms.