int(range) is now stored in a native integer if the range will fit.
Since "large" is stored as INT32 max/min it tends to always fit if the
range is not unbound.

This means that a variable set as int(0..255) will no longer be able
to store values bigger than or smaller than the size of the native
integer, which is an incomatible change, albeit a fairly logical one.