Renamed 'hash' to 'hash_7_8'
Hopefully it will not be used now (this is triggered by yet another
person thinking it's a good string hashing function).

The function is not really suitable for use. The returned values are
higly likely to be the same for similar strings, something that is not
optimal for hashing.

Also added comments about what to use instead (hash_value or one of
the cryptographic hash functions).

It might make sense to add a hash() that actually works, perhaps
siphash?