Made some changes to how the build date and Pike version is resolved.
Instead of having two separate files for version and date (when built with FLAGS_NO_DYNAMIC) there is now one file generated called "constants.js" which is placed in the modref root.
This file is now also used even when the docs are generated without FLAGS_NO_DYNMAIC, so it's always in play.

I also added a "devel" build target in the Makefile, which just copies the css and js in the structure dir to the modref dir.

I also optimized the SVG files.