Compiler [Pike]: Local function declarations are now statements.
This means that they no longer are expressions, and the terminating
semicolon is thus no longer required.

It seems that it was quite complicated to actually use them as
expressions, so the fall-out from the lack of backward compatibility
is likely to be limited, but we may want to consider adding a
corresponding warning to Pike 8.0.