Charset.ISO2022Enc: Fixed symbol name.
The symbol Charset.ISO2022Enc()->charset got renamed to
Charset.ISO2022Enc()->name in the CMOD conversion due to
a mismatch with the name of the field in the old C-struct.

This caused lookup of the symbol "charset" to fail with UNDEFINED,
which in turn caused the internal function transcoder_error() to
generate an illegal NULL T_STRING svalue.

Fixes testsuite failure.