Types: Fixed off by one errors in traversal of the type hash table.
The value pike_type_hash_size is misnamed as it is the mask
(and thus one less than the size) of the type hash table.

This fixes a few loops that missed traversing the last bucket of
the type hash table due to handling pike_type_hash_size as a size.