Invoke signal handlers in the correct process after fork
When using fork() from Pike, the child needs to have its own pipe to
send pending signals over.  If it is shared with the parent, either
the child or the parent may receive the signal.

Close the pipe and create a new one in the child after a fork.  Do the
same for the process wait info.