[build] Explicitely check --with-mpi in sub configure.in files
The ac_user_opts related variables do not always get propagated in the
way that src/modules/configure.in expects. Checking --with-mpi
explicitely in sub-configures that use the setting gets around the
issue.

AC_ARG_WITH lowercases the variable name as well, so
src/modules/configure.in now also lowercases ${with_{$MODNAME}} when
checking whether to build a module.