JSON, JSON5: Regenerate with ragel 7
ragel 6 creates switch statements full of fall-throughs without
properly marking them up as intentional.  This causes tons of warnings
with gcc 7.3.  There is a patch submitted to ragel to add the
annotations, but no release containing it has yet surfaced.
Meanwhile, ragel 7 generates code which does not have the problem in
the first place.  Require use of ragel 7 for now.