__resolve_import can now return an array of source and absolute path.
This is only useful if the Sass/SCSS files resides in a non-standard filesystem. If __resolve_import is overloaded you can resolve the absolute path of the imported file and return the source and absolute path for libsass to keep track of the import chain.
This is needed to resolve nested imports (imports in imported files) if libsass can't access the filesystem.