Refactored how the import handler is defined and resolved.
Instead of assigning the callback to a variable we now do a symbol/identifier lookup for `handle_sass_import` in the current object.

This also got rid of the cyclic dependency wich rendered the wrapper class in the pmod obsolete.

We should probaly add some more checking so that the callback actually is callable and so forth.