Tools.Standalone.precompile: handle args < 0 in PIKEFUN
The 'args' variable is signed even though it is never negative. This
situation is confusing for the C compiler because negative values
are not usually handled. In order to improve this situation the
CMOD precompiler will now check that args is non-negative and either
fatal (in PIKE_DEBUG) or use __buildin_assume to allow the compiler
to optimize away all code path which assume args to be negative.