Make werror() (and by extension exit()) respect current locale
This makes it possible to output non-ASCII text and still have it
legible in all environments.  write() still outputs binary data,
and so will werror() and exit() in 8.0 compat mode.

Also, restore master.pike.in to ISO-8859-1 encoding, which is the only
one supported for the master, and remove the (now superfluous and)
aribitrary use of string_to_utf8() for the copyright message.