Mapping: m1 & m2 is now much faster in cases m2 >> m1
Previously, when doing m1 & m2, we would always copy m2 and delete all
entries from it not present in m1. Now, if m1 has less than half the
amount of entries m2 has, we copy m1, delete entries not present in m2,
and copy the values from m2 for all entries in m1.