Calendar: Added script to (re-)generate the time zone expert system.
The expert system has been updated a single time (2008-01-13)
since it was implemented on 2000-10-01, and it seems to miss
some recent time zones, as well as guess wrong.

It is unknown how the decision points in original expert system
were selected (by hand?), so the new ones differ quite a bit.

Fixes some of [Pike-174].