Master: Re-hook global constants on replace_master().
replace_master() sets the variable is_pike_master in the object
that is to be the new master to 1. Catch this with a setter so
that global functions implemented by the master are actually
implemented by the active master.

This mostly affects the test suite, which replaces and restores
master objects several times.