Build: Fix parallel make
Since module_objects are built in sub-makes, also build
post_modules_objects and pike-complete-stamp in a submake to ensure
correct ordering and that dependencies do not get built while a
submake is also building them at the same time.

There is no need to tell the submake explicitly to build
post_module_objects, since pike-complete-stamp depends on
post-modules-stamp.

Also, join foo_module_objects and foo-modules-stamp into the same rule
since they have the same prerequisites and the same action so thay are
the same anyway (but could accidentally be built in parallel if they
are separate rules).