Locale: Reintroduce deprecated Charset alias
This reverts commit f64cba1ff832383ac7a46b4bbe8a0264700de976.
This reverts commit ffedb3787383db78beddce529eda0bde7e24d40b.

In Pike 8.0.* and below, no deprecation warnings were issued by the
compiler when accessing Locale.Charset, as the deprecated status of
Locale.Charset was hidden from the compiler by a joinnode. In 8.1, the
master will now issue deprecation warnings in such cases. Hence,
re-adding the deprecated Locale.Charset alias seems like a good idea for
users to have a good chance of noticing the deprecated status with an
8.1 release etc. before final removal of the alias in a future release.