Tools.Standalone.precompile: Fix storage offset handling
Programs using their parent may store parent information in the
beginning of ->storage, thus we need to add
p->inherits[0].storage_offset when trying to reach an object's storage.
The compiler/runtime correctly takes care of this in THIS, but the OBJ2_
and GET_..._STORAGE macros did not take this into account.