[debugger] Add WIP code for the debugger.
This commit contains changes in the language core necessary for
enabling the debugger's breakpoint and stepping feature.

It also contains a work-in-progress pmod, Debug/Debugger, that is an
interface between the interpreter and debugging client.

The debugger needs Pike to be compiled '--with-debug'. The core's
behavior remains unchanged in case of this flag's absence.

Co-authored-by: Henrik Grubbström (Grubba) <grubba@grubba.org>
Co-authored-by: William Welliver <william@welliver.org>