Runtime: Modified struct pike_frame.
The field save_sp_offset (which was an offset from locals) is replaced
by a proper pointer save_sp. This is in preparation to allow for the
locals pointer pointing somewhere else than the stack. This also fixes
a potential issue where the interpreter might crash on some cases of
very deep recursion.