Resolve resource leak in Protocols.HTTP.Promise.
Most visible when using https:// URLs to servers that use Keep-Alive,
this resource leak results in an accumulation of defunct file descriptors
for sockets that should have been closed but weren't. The solution is in
two parts:

1) After a Request sends its response, dispose of itself, so the connection
   is returned to the pool
2) In the mini-Session used for Protocols.HTTP.Promise.get_url and friends,
   disable connection reuse.

There appears to be a reference loop somewhere in the Session that prevents
its disposal. Still investigating.